feat(client): add traits representing HTTP requests
HTTP requests can be in two different states (Fresh and Streaming).
Those states are now represented by two different traits. Additionally,
both of those traits require that the `HttpRequest` trait also be
implemented, which provides some common methods that a request can
support, regardless of the state in which it is found.