refactor(client): use a request factory
The RequestBuilder no longer constructs its instance of a fresh request
by directly calling the `Reques` struct's constructor, but rather
delegates to the client's request factory instance.

This has now decoupled the RequestBuilder from any particular request
type, as it uses only methods defined on the `FreshHttpRequest` and
`StreamingHttpRequest` and the instance is provided by the client that
it references.