feat(client): use a factory trait object
This means that the client is decoupled from the particular
implementation of requests and responses.

BREAKING CHANGE: The change removes the `with_connector` public method
of the `Client` (note: it could stay and default to creating an HTTP/1.1
request factory with that connector for the sake of backcompat).
Additionally, the return type of the `send` method of `RequestBuilder`
is now `hyper::client::response::HttpResponse`, rather than `Response`.
The struct has the same public fields and methods, but is still
different in that it does not have a generic type parameter. Even
renaming it to `Response` would still be a breaking change due to this.